Code cleanup

+++ HoverTank1989/NPCVehicle.h
@@ -441,7 +441,7 @@ public:
     DirectX::SimpleMath::Vector3 GetVelocity() const { return m_vehicleStruct00.vehicleData.q.velocity; };
     VehicleData GetVehicleData() { return m_vehicleStruct00.vehicleData; };
 
-    float GetTestTimer() const { return m_testTimer; };
+    float GetTestTimer() const { return static_cast<float>(m_testTimer); };
     void PushAvoidanceTarget(DirectX::SimpleMath::Vector3 aAvoidancePos, NPCVehicle const* aVehicle) { m_npcAI->PushAiAvoidanceTarget(aAvoidancePos, aVehicle); };
     void PushImpactForce(Utility::ImpactForce aImpact) { m_vehicleStruct00.vehicleData.impactForceVec.push_back(aImpact); };
     void PushImpactTorque(Utility::Torque aTorque) { m_vehicleStruct00.vehicleData.impactTorqueVec.push_back(aTorque); };
